+++ setup/dashboard/index.php
@@ -43,8 +43,8 @@ $sysInterfaceInfo = SystemInfo::enuminterface();
                     <?php for ($i = 0; $i < count($sysInterfaceInfo); ++$i) { ?>
                     <tr>
                       <td style="font-size:14px;font-weight:bold;padding: 2px 2px 2px 12px"><?php echo $sysInterfaceInfo[$i]; ?></td>
-                      <td style="font-size:11px;padding: 2px 2px 2px 12px"><span class="text-success"><span id="NetOutSpeed<?php echo $i; ?>">0B/s</span></span></td>
-                      <td style="font-size:11px;padding: 2px 2px 2px 12px"><span class="text-primary"><span id="NetInputSpeed<?php echo $i; ?>">0B/s</span></span></td>
+                      <td style="font-size:11px;padding: 2px 2px 2px 12px"><span class="text-success"><span id="net_<?php echo $sysInterfaceInfo[$i]; ?>_tx">0B/s</span></span></td>
+                      <td style="font-size:11px;padding: 2px 2px 2px 12px"><span class="text-primary"><span id="net_<?php echo $sysInterfaceInfo[$i]; ?>_rx">0B/s</span></span></td>
                     </tr>
                     <?php } ?>
                   </tbody>
@@ -129,12 +129,12 @@ $sysInterfaceInfo = SystemInfo::enuminterface();
                 title="<?php echo T('CURRENT_VERSIONS_CHANGELOG'); ?>"
                 data-placement="top" class="label label-primary tooltips"
                 style="font-size:10px; padding-top:0; padding-bottom:0px; top: -2px; position: relative;"
-                target="_blank" rel="noopener">
+                target="_blank" rel="noopenner noreferrer">
                   QuickBox :: <span style="color: #fff;text-shadow: 0px 0px 6px #fff;"><?php echo "{$version}"; ?></span>
               </a>
             </h4>
           </div>
-          <div class="panel-body ps-container" style="max-height: 350px; padding: 0;">
+          <div class="panel-body ps" style="max-height: 350px; padding: 0;">
             <?php
               $current_commit = trim((string) shell_exec('git -C /etc/QuickBox/ rev-parse --short HEAD'));
 ?>
@@ -142,10 +142,10 @@ $sysInterfaceInfo = SystemInfo::enuminterface();
             <div class="alert alert-default" style="margin-bottom: 4px; text-align: center" role="alert">
                 <?php echo T('QUICKBOX_COMMIT'); ?>
                 <code>
-                  <a href="https://github.com/amefs/quickbox-lite/commit/<?php echo $current_commit; ?>" target="_blank" title="<?php echo T('CURRENT_COMMIT'); ?>" data-placement="top" class="tooltips">
+                  <a href="https://github.com/amefs/quickbox-lite/commit/<?php echo $current_commit; ?>" target="_blank" rel="noopenner noreferrer" title="<?php echo T('CURRENT_COMMIT'); ?>" data-placement="top" class="tooltips">
                     <?php echo $current_commit; ?>
                   </a>...
-                  <a href="https://github.com/amefs/quickbox-lite/compare/<?php echo $current_commit; ?>...<?php echo $branch; ?>" target="_blank" title="<?php echo T('COMPARE_COMMITS'); ?>" data-placement="top" class="tooltips">
+                  <a href="https://github.com/amefs/quickbox-lite/compare/<?php echo $current_commit; ?>...<?php echo $branch; ?>" target="_blank" rel="noopenner noreferrer" title="<?php echo T('COMPARE_COMMITS'); ?>" data-placement="top" class="tooltips">
                     <?php echo T('LATEST_COMMIT'); ?>
                   </a>
                 </code><br>